modificacion en el carrito

+++ Sexshop_TutsiPop/Sexshop_TutsiPop/Models/ProductosCarritoInfo.cs
@@ -9,7 +9,7 @@ namespace Sexshop_TutsiPop.Models
         public string UsuarioNombre { get; set; }      // Nombre del usuario
         public string ProductoNombre { get; set; }     // Nombre del producto
         public int Cantidad { get; set; }  
-        public int Precio { get; set; } // Cantidad de productos en el carrito
+        public decimal Precio { get; set; } // Cantidad de productos en el carrito
         public DateTime Fecha { get; set; }            // Fecha de la adición al carrito
     }
 }
